"Fix" linting error by commenting out new lint failures
Summary:
I'm not sure why this is causing "ant lint" to suddenly start choking, but we should be able
to lint cleanly. Since we're planning on revivifying the IJ plugin at some point, leaving the
code commented out so it's easy to figure out what needs to be done.

Test Plan: ant lint