Fix local classes in JavaFileParser.
Summary:
Java classes can be defined inside of methods, and these local classes live in
a different part of the AST than regular classes and enums. JavaFileParser was
choking on them. Detect this case and skip it, since these classes are not
externally visible.

Test Plan: New unit test in JavaFileParser.