Remove methods from MorePaths that gave us the impression that Buck worked on Windows.
Summary:
Here is the list of methods that I removed from `MorePaths` in this diff:

  public static Path newPathInstance(String path);
  public static Path newPathInstance(File file);
  public static Path separatorsToUnix(String path);

These methods were bad because they purported to return `Path`s with UNIX seprators.
However, the seprator type is not a property of a `Path`. The separator is only
realized when `Path.toString()` is invoked. Therefore, the only way to pass around
a path with a specific separator type is to use a string.

Strings are convenient test values in unit tests and other places, so in the rare
event where separator-type-consistency is merited, the following methods from
`MorePaths` should be used:

  public static String pathWithUnixSeparators(String path);
  public static String pathWithUnixSeparators(Path path);

Test Plan:
Sandcastle builds.
After we export this change to GitHub, I'll see how things stand with
Buck on Windows on my home computer.