Build the user specified targets, not just the ones w/o incoming edges.
Summary:
I am working on some changes that could result in a disconnected
`DependencyGraph` such that some of the connected subgraphs should not be built.

Currently, we build all nodes in the graph with no incoming edges, which has the
side effect of building all nodes in the graph.

This diff changes things so that we only build the rules (and their transitive deps)
specified by the user. This way, disconnected subgraphs will not be built.

Test Plan: Sandcastle builds.