Make it possible to override _all_ the deps for a BuildRule.
Summary:
Previously, `getEnhancedDeps()` provided this ability, but things have evolved
(see https://phabricator.fb.com/D1259127) to where there are "extra deps"
(that could come from things such as SourcePaths) that are not removed from
the deps, regardless of what `getEnhancedDeps()` does.

This is a smart default, as it prevents us from shooting ourselves in the foot,
but I need more flexibility in order to implement `gwt_application()` efficiently.
To underscore the danger, the method I am introducing that lets you override
all of the deps is named `iKnowWhatIAmDoingAndIWillSpecifyAllTheDepsMyself()`.

Test Plan: Updated existing test in `DescribedRuleBuilderTest` and added another test to it.