Remove scheme gen from ProjectGenerator
Summary: ProjectGenerator does too much. It generates both separate and
combined project styles as well as schemes and workspaces. My plan in
this stack of diffs is to pull scheme and workspace generation into
their own pieces entirely and to restrict ProjectGenerator to the
separate project style while removing the half-implemented combined
style. This will leave us with both a functional way to generate whole
iOS app configurations without Xcode UI and it will leave us in a better
place to extend ProjectGenerator to generate other styles of projects
later if we want to bring the combined project back.

Test Plan: waitforbuildbot