Make the ErrorListener for a BuckEventBus used in tests less chatty.
Summary:
I changed the `ErrorListener` to check the level of the message that it
prints so that it should actually only make noise when there is an error.

I believe that I originally introduced `ErrorListener` in
https://phabricator.fb.com/D1067125 because, in the course of debugging
a unit test, I was missing error messages from Buck that would have been
helpful. This should preserve that property.

Test Plan:
Ran `DescribedRuleTest` and nothing was printed to stderr.
Previously, this would print "cheese" because it executed an `EchoStep`,
which is implemented as:

```
  context.getBuckEventBus().post(LogEvent.info(message));
```