Make None the default value in buck.py for a string or number.
Summary:
This will make it easier to implement a field of a `ConstructorArg`
of type `Optional<Integer>` whose value will now be `Optional.absent()`
when the field is unspecified. This will make it possible for the following
to work correctly:

```
  int value = my_field.or(9);
```

Without this change, `value` will be `9`, even if this is what was specified
in the `BUCK` file:

```
  my_field = 0,
```

The old logic assumed that not specifying a value for `my_field` and
specifying a value of `0` were equivalent, so both were turned into
`Optional.absent()`.

Test Plan:
Added a test case to `DescribedRuleTest`.
Updated unit tests: all pass.