Introduce BuildableParams to hide information from BuildRuleParams.
Summary:
This makes it easier to instantiate a `DefaultJavaLibrary` or `PrebuiltJar`.

As a follow-up, I will try to change `Description.createBuildable()` to take
a `BuildableParams` instead of a `BuildRuleParams`; however, that is a bit
messy because there are some `Description`s that depend on the additional
behavior provided by `BuildRuleParams`.

Test Plan: Sandcastle builds.