Ensure genaidl rules generate an output.
Summary:
We could determine the output file name for an invocation of genaidl, but it's simpler
to just produce a source zip with a file name of the format that the javac steps can
make use.

This moves us one step closer to being able to remove genfile().

Test Plan: buck test --all