Prevent buck.py from printing pipe errors on exit.
Summary:
We were seeing test failures on Linux that bisected to D1243758. It's not clear
why that diff started the failures, but what's happening is that sometimes
buck.py prints to stderr when it closes, and some tests assert that nothing is
printed to stderr. The underlying cause is that flushing a file descriptor can
fail if that descriptor is actually a pipe whose other end is already closed.
Suppress the error.

According to https://mail.python.org/pipermail/tutor/2012-April/089278.html,
this stderr output doesn't happen in Python on Mac.

Test Plan:
Maddeningly, the test failure only shows up in `buck test --all`.
Run that in a loop ten times to be absolutely sure it doesn't fail again.
Revert this diff and watch the failures come back.