Call JavacErrorParser from JavacInMemoryStep.
Summary:
4b46be63742ada5c2e0ac5c504e70361c2c103e1 created a listener in
`BuildCommand` to collect missing symbols during the build.
Now that `JavacErrorParser` is available, parse compiler errors
and post these symbols to the EventBus.

Test Plan:
For now just setting a breakpoint to confirm that symbols are making
it. This will be tested as part of the integration tests of the autodeps
feature.