Change gwt_binary() to output a .zip file.
Summary:
As David Ostrovsky pointed out on https://github.com/facebook/buck/issues/109,
the output from the GWT compiler is a zip that does not have the standard .war
file structure (even though the argument to the GWT compiler is `-war`), so it
is more appropriate for the output to be a `.zip` file.

Test Plan: Retrofit an existing GWT application with `BUCK` build files and build it using Buck.