Bash escape the empty string to "''".
Summary:
When running `buck test --all -v 2`, I noticed that the command printed
to the console was not copy-pasteable. Because the second argument to
`com.facebook.buck.junit.JUnitRunner` is often the empty string, it was
was not being printed to the console as part of the command, which is
why it could not be copy-pasted.

Test Plan:
Updated `EscaperTest`.
Copy-pasted the `java` command to run a test as produced by
`buck test --all -v 2` and verified that it worked.