Output a useful error message if a genrule relies on a location that's null.
Summary:
When using a genrule, it's possible to use the "$(location :target)" construction in
order to get the output of a build rule. This is fine, unless that rule happens to
have no output. In this case, throw a meaningful error message rather than an NPE.

Test Plan:
Add a test
buck test --all