[ios] Allow project to specify its own asset catalog build script
Summary:
@bhiller noted that using asset catalogs would require engineers to run buck
project upon checkout before being able to build. At this time, this is not
necessary in fbobjc and should remain unnecessary until all of the repository is
converted.

Test Plan:
Add .buckjavaargs to set the new system property in fbobjc buck invocations and
see projects using apple_asset_catalog create targets with a reference to the
custom build script.