[ios] Fix build phase generation when using custom asset compilation script
Summary:
The previous change to allow a custom asset catalog compilation script was not
tested properly. Buck still placed its own compilation script in an output
directory. The reference to the script in the project was also incorrect.

Test Plan:
Migrate a project to use apple_asset_catalog. Verify buck does not place its own
compilation script anywhere. Build the resulting project and verify the assets
are placed in the correct place.