Re-record DX's "find-usages" test
Summary:
It looks like the output was made more verbose, but the test wasn't
updated.  It's hard to tell exactly where the change was made.  The test
was broken in a different way in 6bf7f30e6968d991 because it wasn't
flushing output.  The flushing was fixed in 1c7c32e2ff5d9f6a, by which
time the output had changed.

Test Plan: ant dx-tests