Pass the correct client environment to DefaultPropertyFinder.
Summary:
In order to make the use of the Buck daemon transparent, Buck needs to
use the process environment when running as a single process and the client
environment sent by Nailgun when running as a daemon. These changes extend
DefaultPropertyFinder to accept an environment when it is constructed and pass
the correct environment to instances when they are created.

Test Plan: buck test --all