Use in-process dexing for pre-dex and dex merge
Summary:
My previous diff laid out an argument that this is safe.  This diff
actually integrates it.

A DxStep.Option controls whether this is enabled or not.  We only turn
it on for pre-dexing and dex merging.  Release builds will still use
stock DX.

DxStep is always a shell step, even though it doesn't always use the
shell.  I think this is okay.  Splitting it into two different classes
requires changing a lot more code.

Test Plan:
Built FB4A with and without this.  All of the generated classes.dex
files (which are mostly inside jars), including the merged ones, are
identical.