Disable test that fails when inotify watch limit reached.
Summary:
The DaemonIntegrationTest is run on the CI servers, which also have a bunch of other work
happening on them at the same time. This results in us bumping into the limit of inotify
watches, causing the CI builds to constantly fail. This reduces their value to us.

Temporarily disabling the failing test until we get a fix in place.

Test Plan: CI server presubmit run.