Add support for sending input to a ShellStep
Summary:
Some shell steps can execute commands which expect some input via
stdin, which was currently supported in the ##ShellStep## class.
This diff adds a ##getStdin## method to this abstract class that
returns an optional string.  By default, this is absent, but the
developer can override this method to write stdin into the shell
command.

Test Plan: - ant test