Ensure buck, and its tests, pass cleanly on Java 8.
Summary:
Looks like Java 8 rejigs the way that Sets are ordered by default. Get our tests green again,
but I suspect we need to audit where we add lists to rule keys and ensure that they're always
sorted.

When Buck is compiled using Ant, the source and target versions of Java are not specified, so
whatever is being used by the system is selected. Android, OTOH, popularly targets Java 6,
so we need to ensure that the correct versions are being tested in the `DalvikStatsToolTest`.

Test Plan: buck test --all on Java 7 and 8