Introduce AndroidPackageable interface to collect artifacts that go into an APK.
Summary:
An `AndroidPackageable` decides what artifacts should end up in an APK as well
as other dependencies that need to go with it. The motivation behind doing this
is android_resource can remove its runtime dependency on android libraries while
at the same time ensuring that those dependencies end up in the APK.

This essentially distributes the logic in `AndroidTransitiveDependencyGraph` down
to the individual build rules. `AndroidPackageableCollection` is the union of
`AndroidTransitiveDependencies` and `AndroidDexTransitiveDependencies`.

Test Plan:
buck test --all
buck build instagram