Semi-unify asset and non-asset native libraries
Summary:
Previously, native-libraries-as-assets were added to the apk in the
aapt_package step, while normal native libraries were added during
apkbuilder.  This probably came about because "most" assets are
associated with android_resource rules, which are handled by aapt (which
knows how to bundle assets), and native libraries are supported directly
by apkbuilder.

However, this makes no sense structurally.  Asset and Non-asset native
libraries are 100% identical except for "assets/" being prepended to the
in-apk path.  So let's have them both be handled by (mostly) the same
codepath in apkbuilder.

The motivation for this is the elimination of
AndroidTransitiveDependencies.nativeTargetsWithAssets.  We're planning
to refactor how Android transtive dependencies are handled, and this
thing is kind of getting in the way because its the only thing that's
not a simple Set<Path>.

If we want to optimize exopackage abi calculation in the future, we
might need to add back something like this for both asset and non-asset
native libs, but I still think it's worth deleting it now to simplify
the refactor.

Test Plan: Integration tests.