Introduce a Repository abstraction.
Summary:
One of the planned changes to Buck is the introduction of "cross-repo targets". These are
targets that allow building of targets from other (buck-ified) projects that are checked
out "elsewhere" on the filesystem. As an example, perhaps you have a suite of common
library functions checked out under "/src/common" and your project under "/src/project".
A cross-repo target would allow a target in "/src/project" to refer to a target in
"/src/common".

This introduces a Repository abstraction upon which we can start iterating the
implementation of these targets. I suspect we're missing a "Repositories" abstraction too,
but baby steps.

Test Plan: buck test --all