Add Apache's commons-compress library to third-party
Summary:
This provides a zip archive extractor implementation which supports
some features that the built in one does not, namely support for
external attributes so we can preserve execute permissions in the
artifact cache.

(Note: this will and should *not* be used to replace our internal
implemenation for creating the zip archive).

Test Plan:
Ran builds with subsequent diffs that use this implemenatation in
`Unzip`.