JavaDirectoryStep: only reset compressed size for deflated entries
Summary:
D984455 changed the ZipDirectoryStep to work around a rare bug which
caused zip exceptions when re-compressing the entry changed the
compressed size.  However, we should only do this for "deflated"
zip entries, as "stored" zip entries require the compressed size
to be equal to the uncompressed size (as documented at
http://docs.oracle.com/javase/7/docs/api/java/util/zip/ZipEntry.html#getCompressedSize()).

GitHub: fixes #132

Test Plan:
Built a java binary dependening on a prebuilt JAR with an embedded
JAR "stored" entry.