Easy: Buck iOS tests: Ensure deps is non-null in tests
Summary:
I'm about to make it so we depend on the 'deps' field of
IosLibrary etc. being non-null. This is always true in production,
but in tests, we manually create the IosLibraryDescription.Arg
value and often leave fields null in tests.

This makes sure the deps field is always non-null in our tests.

Test Plan: ant test