Make autodeps support missing symbols in the current package.
Summary:
When a java package is divided up into multiple java_library rules, it's
possible for those rules to be missing deps on one another. In that case, there
will be no package name given with the missing symbol in the compiler error
(because java allows different files in the same package to reference each
other's symbols unqualified). We need to infer the package in these cases.

This diff uses `JavaPackageFinder` along with the path of the source file to
infer the package. Here we attach the package finder to
`ExecutionContext` (as opposed to a previous approach where we put it
directly in `JavacStep`). The jury is still out on which one we prefer.

Test Plan: New unit tests for `JavacErrorParser`.