Handle nullable exception messages cleanly in BuckEventAndroidLogger.
Summary:
The Android classes appear to use @NonNull when a parameter can't be null, which is the exact
opposite way that the rest of the Buck code deals with these things. According to the docs,
both the `throwable` and `errorFormat` parameters to Android's `ILogger.error()` method are
optional.

Test Plan:
buck test --all
Visual inspection.