Write Xcode projects, workspaces, and header maps as read-only
Summary:
We're soon going to remove the checked-in Xcode projects
from fbobjc.

This diff uses the new functionality from D1401826 to write
read-only Xcode projects, workspaces, and header maps.

I confirmed the Xcode UI handles this gracefully; it prompts
the user to let them know the project file is not writable
(it says "locked") if they try to change it.

Depends On D1401826

Test Plan: Integration test included.