SortedSetTypeCoercer: throw error on duplicate elements
Summary:
Previously, the ##SortedSetTypeCoercer## class would silently throw
away duplicates when converting from listy types.  This diff changes
this behavior so it throws a coersion error instead.

This probably isn't a huge deal for how we currently use this class,
as it's typically used to convert from a list of homogenous types.
However, when used with a collection of heterogenous types, where
some objects of different types are determined to be equal by the
comparison function, then this will help catch weird cases where
one of the objects is silently overriden (e.g. for the case of a
"named" source path, where a source path derived from a listed source
has the same "name" as a source path derived from a build target
representing a genrule).

Test Plan: added a unittest