NOTESTS status for tests with no output
Summary:
Buck's console output aggregates test results and summarizes them with a
single word, which by default is "PASS".  I find that's really confusing
as the test isn't really passing -- no test ran!  I find printing
"NOTESTS" is more informative, especially for the cases where it's not
obvious a test has been `@Ignore`'d.

(An alternative would be to not output any information for classes that
don't have any tests in them, but that didn't feel right -- it's better
to have a positive signal that nothing ran instead of absence-of-output
to indicate nothing ran.  The latter just looks and feels like a bug.)

Test Plan: Run with an empty test class!