Make test that check output line containing running time data more resilent.
Summary:
This diff introduces a RegexMatcher helper that can be used in unit test assertions to check whether a string (or a part of it) matches some regular expression.
As an example AssumptionViolationsTest is migrated from basic string comparision to a regex based one. Since the test was comparing buck output log containing the execution time string on it, this diff is suppose to make this test more resillent.
Since we have ~4 more junit classes that do this kind of comparison I will migrate those classes to use RegexMatcher in following diff.

Test Plan:
1) AssumptionViolationsTest should pass: ant java-test -Dtest.class=AssumptionViolationsTest
2) Modify pattern in AssumptionViolationTest in a way it fails, then run the test again
3) Revert changes from (2)