Use /data/local/tmp for storing exopackage jars
Summary:
Previously, we stored exopackage jars in the app's private data
directory.  This had 2 main benefits: (1) it closely mirrored how we
handle secondary dex files in production, and (2) the files were
automatically cleaned up when the app was uninstalled.

Unfortunately, it required a working "run-as" command on the device,
which turns out to be less reliable than I expected.  So we're switching
to an alternate mechanism: storing jars in /data/local/tmp.  It's
possible that this directory might not be shell-writable on some
devices, but our recent experiments suggest that it will be more
reliable than run-as.

These files will not be automatically cleaned up on uninstallation.  To
mitigate this, I try to delete them in "buck uninstall" which I will be
encouraging people to use.

We could use "adb push" to install these files, rather than relying on
the agent, but it looks like that will require 2 additional adb round
trips to avoid creating a world-readable file (see
https://github.com/android/platform_system_core/blob/android-4.4.4_r1/adb/file_sync_service.c#L309),
and that could hurt performance, so I'm not making that change now.

Test Plan:
Did an install and uninstall.  Saw dex files installed and deleted.  Was
able to load them.

Note that this will need to be synchronized with a change in fbandroid.