Copy local.properties if it exists
Summary:
This change propagates the project's local.properties file to any temporary
test projects it creates. This solves the issue where if the SDK or NDK paths
were specified in a project's local.properties file instead of env variables,
the temp test projects generated failed to build correctly.

Test Plan:
1. Check out buck
2. Create a local.properties file in the root directory of the buck project. Use it to specify an SDK and NDK. For example:

    sdk.dir=/path/to/android-sdk-macosx
    ndk.dir=/path/to/android-ndk-r9b

3. Make sure the ANDROID_NDK env variable isn't set (e.g. run `unset ANDROID_NDK`)
4. Run the android tests: `buck test //test/com/facebook/buck/android`

Expected results:
The tests pass. (Prior to this change, they failed with an error that the NDK couldn't be found.)