Make some more coercion errors more human-readable
Summary:
A few more container type coercers were suppressing inner coercion
errors containing actionable information for the user, replacing them
with more obscure higher-level errors messages (e.g. "cannot coerce
to ImmutableMap" instead of "no such file or dir: 'foo/bar.java'").
This diff fixes this in the remaining container types and adds tests
to verify this behavior.

The tricky case in all this is the `Either` type.  Previously, the
`Either` coercer would first try the left side, and if that failed
fallback to the right side, throwing whatever error the right side
produced or returning its result.  This made it very difficult to
propagate the correct error message to the user.  For example, when
coercing a list of paths using `Either<List<SourcePath>, SourcePath>`,
if one of the paths in the input list didn't exist, the coercer
would fall back to using the right side coercer, which would fail
and propagate an error message explaining to the user that it couldn't
coerce a list type to a path type.

To fix this, this diff adds code to the `Either` coercer class to
classify the input object and left/right coercers (i.e as either
a collection, map, or other), so it can pick the "correct" coercer
and throw it's corresponding error message on failure.  Long-term
we should probably come up with something more robust, such as extending
the coercer API to list the input object types it can parse.

Test Plan:
- included unittests
- https://www.facebook.com/groups/560619233986696/permalink/703393513042600/