Fix the hash that works with Gradle in exopackage.html docs.
Summary:
https://twitter.com/hans_d/status/492021893026021376 rightfully pointed
out that the revision in our documentation does not build out of the box.

This is because AntennaPod requires you to create a config file:

https://github.com/danieloeh/AntennaPod/wiki/Building-AntennaPod#step-2-create-flattrconfigjava

We included a commit to create a dummy config so that AntennaPod builds out of the box:

https://github.com/dreiss/AntennaPod/commit/c2080b1dfd17fc371e04ce1e7b39ebadaf3cb7a7

But we forgot to use it as the base diff in our instructions (instead, we
were using the base of our fork, b2ae4c15d13840619fb09cc141b290449e40a6a4).

I verified that the build instructions in our docs work now. If there is a failure, you
likely need to run `$ANDROID_HOME/tools/android sdk` and install the
appropriate packages, which may include:

* Android SDK Build-tools 19.0.3
* Android Support Repository

Test Plan: Followed the updated build instructions and verified that they worked.