Ensure values_file is built before android_build_config().
Summary:
Fixed this bug by adding `values_file` to the `extraDeps` of the
`BuildRuleParams`. I think this is an example of a bigger issue that
@simons is looking at where `SourcePath`s are not automatically
added as `extraDeps` to a `BuildRuleParams`.

Note that this also fixes the issue for `build_config_values_file` in
`android_binary()` because `AndroidBinaryGraphEnhancer` delegates
to `AndroidBuildConfigDescription.createBuildRule()`.

Test Plan:
Built an `android_binary()` that had a transitive dependency on an
`android_build_config()` with a `values_file` argument. Verified that
the APK did not built without this change, but built with it.