Allow Descriptions to declare which Flavors they support.
Summary:
As we build the target graph, it's useful to know whether the targets being referred to
can actually be built. Currently, the only flavour that can be addressed directly is the
default one. As we expose Flavors directly, however, this will not be the case.

Test Plan: buck test --all