Fix checking if git repo has changes.
Summary:
Checking the local repo for changes with the --short option is flawed, as the user might have
changed his git status output, e.g. to:
[status]
	submodulesummary = -1
	short = YES
	branch = YES
in which case the branch is also visible with the --short option. Use --porcelain instead:
"Give the output in an easy-to-parse format for scripts. This is similar to the short output,
but will remain stable across Git versions and regardless of user configuration."

Test Plan:
Change repo, make sure buck thinks it's changed.
Adjust .gitconfig to above listed one, make sure buck doesn't think it has changes.