Enforce Eclipse warnings by using ecj as a compiler.
Summary:
Update the Ant build to use the Eclipse Compiler for Java (ecj) with strict
checks enabled so that we can catch issues at build time, such as missing
`@Override` tags. Up until now, I have been manually fixing these errors as
I discover them when I load newer versions of Buck in Eclipse (most Buck
team members use IntelliJ and are unaware of these violations). It is finally
time to replace my manual labor with a computer.

The set of checks performed by ecj in our build are listed in
`third-party/java/eclipse/org.eclipse.jdt.core.prefs`. This is only a subset of
the available options we could configure, as the full list is at
http://help.eclipse.org/kepler/index.jsp?topic=%2Forg.eclipse.jdt.doc.isv%2Freference%2Fapi%2Forg%2Feclipse%2Fjdt%2Fcore%2FJavaCore.html.

My immediate goal in this diff is to curtail my current manual labor. As such,
it was more straightforward to hook ecj into Buck's Ant build than to introduce
the option of building with ecj in Buck itself. Admittedly, this makes a bit of
a mess in `build.xml`, but `build.xml` is already pretty lousy to begin with.

Giving Buck users the option to build with ecj and their own ecj configuration
would be a good thing to do in the future. In particular, if we could rely on Buck
to use ecj when building itself, then we could stop running ecj as part of Ant,
which would simplify `build.xml` considerably.

The changes to `build.xml` in this diff require users to upgrade to Ant 1.8 or later.
See the docs on `compilerclasspath` in http://ant.apache.org/manual/Tasks/javac.html.

The biggest change in this diff is to the `compile` target in `build.xml`. Specifically, `compile`'s
lone `<javac>` task is now split into three `<javac>` tasks:

* The first uses ecj to compile the dependencies of `//src/com/facebook/buck/apple/xcode/xcconfig:xcconfig`.
* The second uses javac to compile `//src/com/facebook/buck/apple/xcode/xcconfig:xcconfig`.
* The third uses ecj to compile the rest of the `src/` directory.

We prefer to use ecj for its strict compilation settings, but that is not possible when compiling
`//src/com/facebook/buck/apple/xcode/xcconfig:xcconfig` because it contains generated code
that is full of warnings and is not within our control to fix. This diff also creates an
Ant macro for compiling with ecj (aptly named `<ecj>`) in hopes that we can convert more of
the `<javac>` tasks in our `build.xml` to use ecj. In particular, converting
`<target name="compile-tests">` to use `<ecj>` will be done in a follow-up diff.

Additional notes on changes that needed to be part of this diff:

`jdtCompilerAdapter.jar` was taken from the `org.eclipse.jdt.core_3.9.2.v20140114-1555.jar`
file that is already in the repository. According to the Eclipse docs, extracting
this file is required to use ecj from Ant:
http://help.eclipse.org/luna/index.jsp?topic=%2Forg.eclipse.jdt.doc.user%2Ftasks%2Ftask-ant_javac_adapter.htm

`<compilerarg value="-implicit:class" />` was removed from `build.xml`, which was
originally added when neo4j support was added to Buck. neo4j was later removed for
licensing reasons, so this seems safe to remove from `build.xml`.

Test Plan:
Sandcastle builds.
Introduce Eclipse errors, such as a missing `@Override` tag and try to rebuild with Ant;
verify that this fails the build.